FSP: Fix state check to trigger link down
The fsp_mbx_err state is currently abused in the HIR path and its use there
is incosistent with its intended purpose. So add a fsp_mbx state to indicate
completion of the HIR steps, so the R/R procedure is triggered in the normal
path.

When in the FSP error path, do not bring the PSI link down if the link down
check fails and the fsp mbx state is fsp_mbx_hir_seq_done. Also use this
additional state to determine a late state error.

Tested on P7. Awaiting response to FW issues on P8.

Signed-off-by: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>