DPO: Add support for Delayed Power Off sequence on FSP machines
This enabled DPO sequence handling in Sapphire. Delayed power off notification
comes from FSP when the user requests for it through either ASM interface or system
front operation panel. After receiving a valid DPO init message from the FSP, Sapphire
intimates the host through an OPAL message and then sends an acknowledgement message to
the FSP for the last DPO. The host is required to shutdown all the guests, user space
programs and then shutdown itself. The shutdown sequence of the host will eventually
make the opal call OPAL_CEC_POWER_DOWN receiving which the Sapphire will send a shutdown
command to the FSP. Sapphire has got 45 minutes from the time for the initial DPO message
to this shutdown command to FSP failing which the FSP will assume that the DPO sequence
was successfull both in Sapphire and host, hence force shutdown the system.

Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>