elog: Queue errors to be committed in FSP with timeout value.
This patch aims to queues all the OPAL errors to the FSP
with a time-out value. If the error gets
committed in FSP and an ack is recevied for the same
within the time-out interval, then it is not reported
in the host. Else on expiry of the time-out value the
error is committed in the host.

Currently the time-out value is set to 180 seconds.

Signed-off-by: Deepthi Dharwar <deepthi@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>