elog: Move error log data structres from opal.h to new file errorlog.h
Currently, all the error log data structures needed to commit an error
in PEL are declared in opal.h and are exposed to kernel.
At present we do not have any infrastrucutre in the kernel to make use
of these ABIs.

Any changes to these strucutes in the future can
potentially cause an ABI breakage. In order to avoid
any breakage this patch moves all the error log data structures
to a new errorlog.h file and deprecating the OPAL tokens for
error-log write functionality calls only.

Once the kernel support is added (anytime in future),
one needs to move all of this back to opal.h and re-enable
the OPAL tokens for the same.

Signed-off-by: Deepthi Dharwar <deepthi@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>