chiptod: Improvement to make quicker recovery from TOD errors.
When we get an HMI for TOD errors, chip TOD is generally not in running state.
Before we fix TOD state we need to do a quick check whether it is running or
not. But in current implementation instead of doing a quick check we do a
poll check for TOD running. This delays the process of fixing TOD state and
on a system with more cores per chip we may see soft lockup's even after
successful recovery. This patch fixes this issue by introducing a quick check
for TOD running state and provides quicker/faster recovery from TOD errors.

Signed-off-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>