Initial code for timestamps in log
This is the initial patch for having timestamps in the log.

It currently only wraps prerror to our prlog() function and thus only
(very slightly) modifies bootup log.

we use the timebase as an indication of the progression of time. It is
not perfect, and is indeed reset back to zero during boot, but it should
serve adequately for our needs of "approximately this much time elapsed
between log entries".

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>