Use PR_EMERG priority in assert() codepath
Moving assert_fail() out of libc and into core/utils.c so that we can
sanely call prlog(PR_EMERG).

We shorten it from three fputs calls down to one prlog() call.

This may increase the number of cycles and stack usage for when we
hit an assert, which may not be desirable.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>