epow: Fix mbox command encoding to remove FSP orphan responses
This patch changes the EPOW related mbox commands encoding to
make sure FSP message response structure is allocated. So that
when the command returns from FSP, it can be handled appropriately
and orphan response will not be logged. This also adds some more
in-code documentation explaining why panel status request based
handshaking is required with the FSP.

Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>