epow: Change how the old and the new system epow status are maintained
This patch changes the way how the old and the new system epow status
are maintained. After receiving FSP EPOW async message, only reset
those parts of system EPOW status which will be updated with the
new FSP async message. This patch also fixes a problem with comparing
the contents of the new and the old system EPOW status.

Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>