PCI: Support parallel scanning
Currently, the tasks of scanning PHBs are done on master CPU one
by one. The patch intends to do same tasks on multiple CPUs in
order to save booting time with help of additional flags to PHB.

With the patch applied, we saves 22 seconds the tasks to reset
and scan 8 PHBs on one P8 box from 37 seconds to 15 seconds.

NOTE: the printed logs during PCI enumeration should include
PHB index to be self-explaining enough. I'll fix it later.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>