occ: Increase wait time to 60s for OCC to init
OCC F/w team recommends 60s timeout waiting for OCC to init. OCC has
to wait for memory throttle calibration from hardware procedure and
that takes 10s of seconds on large memory configurations.  In one of
the failing case, it took 24s for OCC to init.

Typically OCC takes 2-5 secs to boot and we do that in parallel with
skiboot inits.  But on certain corner cases with large memory,
we have to wait for 60s before we give up.

Signed-off-by: Vaidyanathan Srinivasan <svaidy@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>