Fix gitid generation when not in a git repository
If skiboot doesn't live in a git repository, the makefile will go
searching up the directory tree for one. To save confusion, only look in
the source tree for the git repository.

The gitid can be overridden by passing GIT_SHA=foo as an argument to
make. This is useful for doing builds outside of a git tree, as
buildroot does.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>