SURV: Trigger HIR on surv sysparam query error
For the case where the survserver on the fsp server is dead (for whatever
reason [1]), even before the first time query via sysparam of the surv
status by sapphire, we get an error response to the sysparam query.

We should apparently trigger a HIR in that case (same as phyp).

[1] survserver has a real bug on a 'fsptelinit --disablerecovery'
followed by a 'kill -9 <survserver_pid>'

Fixes https://bugzilla.linux.ibm.com/show_bug.cgi?id=114646.

Signed-off-by: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Signed-off-by: Jeremy Kerr <jeremy.kerr@au.ibm.com>