ipmi/power: Update chassis control command
Linux sends us a 0 when shutting down. This means we don't need to pass
the u64 to the IPMI driver. Add a check that the value is what we expect
in case Linux changes it's behaviour in the future.

When rebooting, we should send the BMC a HARD_RESET command (0x03), not
POWER_CYCLE (0x02).

While we are here, trim some whitespace and drop opal from the IPMI
function name for readability.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>