ipmi/power: Add power state IPMI command
This sends command 20.7, Set ACPI Power State. It is to be used to
inform a BMC of the runtime state of the system. We drop the ACPI part
from the function name to avoid confusion.

As soon as IPMI is up, the palmetto platform init code will set the
power ptate to S0/working.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>