platform: add a platform hook for loading external resources
Currently, in core/init.c we do a fsp-specific load procedure to grab
the kernel image.

We'd like to do two things: allow other types of resources, and have
paths for non-FSP platforms to perform loads.

This change adds a platform-specific load_resource hook, and moves the
currently loading code to fsp_load_resource. To allow other resource
types, we add an identifier to indicate the type of resouce to load.

Signed-off-by: Jeremy Kerr <jeremy.kerr@au.ibm.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>
Acked-by: Stewart Smith <stewart@linux.vnet.ibm.com>