hw/ipmi: Add API for setting sensors
This allows setting a given IPMI sensor to a given value.  There are
helpers for setting the firmware boot progress that will be used for
updating the BMC with the host boot progress.

The sensor ids are parsed from the device tree. If the sensor cannot
be found we will silently continue.

Signed-off-by: Joel Stanley <joel@jms.id.au>