core: pace the loop in cpu_wait_job
When waiting for a job to complete, users call cpu_wait_job. This causes
the pollers to run and take locks eg. dummy_console takes con_lock.
These locks may also be contended by the job that is running, and this
contention between the pollers and the job stops the job making forward
progress.

The 10us delay was picked as a reasonable compromise.

With this change Palmetto with DD2 PCI probe time goes from 8e10 to 4e8
ticks.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>