phb3/capi: Add two new modes to opal_pci_set_phb_capi_mode
For user initiated capp recovery, provide a mode to turn snoops off.  The perst
alone does not turn snoops off and we need to do this as part of the capp
recovery procedure before reinitializing the phb.

A second mode turns snoops back on after recovery.  The driver needs to do this
after it reinitializes the PSL otherwise tlbies could come in before the psl is
initialized.  Also write 0 to capp error status and control as part of the
recovery procedure.

Put modes as flag defines in opal.h so the driver can pick them up.

Add a dt property "ibm,capi-modes" which tells the driver which modes sapphire
supports.  For backwards compatibility with older opals.  Also, the driver can
disable reset in sysfs if not supported.

Move the mode checking into phb3.c so it's all in one place.

Signed-off-by: Ryan Grimm <grimm@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>