plat/palmetto: Fix SIO UART vs UART setup and iBT setup
The AMI images use the virtual UART, not the SIO UART, so configuring
the SIO the way we do is incorrect. Additionally, they don't configure
the interrupts properly (bad polarity for VUART and bad number for iBT).

This reworks the inits to fix that up:

 - All SIO interrupts are set to level low

 - Check if VUART is enabled. If yes, configure and use it (and disable
SIO UART which hostboot might have left enabled).

 - Else, reconfigure VUART LPC address and IRQ properly

 - Configure iBT LPC address and IRQ properly

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>