ipmi/bt: Rework iBT register access
The bits in the control register are mostly write-1-to-clear, so
the rmw sequences in bt_setmask() and bt_clearmask() don't work.

Additionally, H_BUSY is weird as it's a write-1-to-toggle, so let's
write a "safe" function that sets it to the desired state based on
its previous state. (We can optimize that further later).

Also enable interrupt operations.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>