uart: Give UART it's own OPAL console callbacks instead of dummy console
This means the Linux output no longer gets into our internal log,
which makes dumping of it from Linux a lot nicer. It will also allow
us to improve the way we do the bufferring for Linux and to exploit
eventually the TX interrupts. It will also allow us to implement
some form of timeouts for the OPAL console variant of it so we don't
get stuck of the BMC doesn't consume from the virtual UART.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>