uart: Keep synchronous console flush when IRQ not working
Some boards are missing an EC for interrupts (or might not have
the FPGA properly flashed). In that case the serial interrupt isn't
working.

We attempt to detect this by enabling all interrupts by default and
when the first one occurs, we mark interrupts as "functional".

Until they are detected as such, we keep all output flushes fully
synchronous.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>