bt: Add call to bt_poll to kick start message state machine
The bt interface relies on a fairly simple state machine to send
messages to the bmc. Messages are added to a queue for sending and
something calls the state machine to start sending messages, usually
via opal_poll_events. However under some circumstances this doesn't
happen so we need to crank the state machine to start start sending
messages.

This patch adds a call to bt_poll to make sure we start sending a
message which should ensure the bmc generates an irq once the message
is sent to continue processing.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>