bt: Add locking for bt hardware polling
This patch adds some locking to the polling of the bt
hardware/state. Prior to having proper interrupts and events working
this was not so critical as we would only ever have one call to
bt_poll running at a time. However functioning interrupts mean we can
have multiple calls to bt_poll so we need to add some locking to ensure
consistency between the state machine and the hardware.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>