bt: Fix message timeouts
The initial timeout scheme started a message timeout when it was added
to the queue rather than when transmission was started. To avoid
timing out messages that just got stuck behind a couple of slow ones
we should instead just check that the active message has not timed
out.

This patch ensures timeout values are only checked for the currently
active message.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>