i2c: Improve operations without interrupts
We use a timer for running the state machine without interrupts. When
we get proper timers using the SLW engine, we'll be able to also reduce
the timeouts but for now, use a large (4s) per-char timeout for polled
mode as it depends on the opal poll frequency which can be very slow.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>