i2c: Update DT bindings and minor changes
The i2c bus frequency is now per port and we use more standard properties
this and the clock frequency of the master cell itself.

Also use "ibm,opal-i2c" as a compatible match to indicate that the DT node
is compatible with the OPAL APIs, whether it's a P8 i2c bus or something
else (we will eventually expose the centaur i2c's too for example).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>