PHB3: wait 1s after PERST deassert for CAPP adapters
On a re-ipl or warm reboot, Sapphire asserts and deasserts PERST to each slot.
This results in the fpga image loaded into the flash for a CAPP adapter.  HMIs
have been observed with a 200ms wait following PERST deassert, so bump time up
to 1s.  Do this for all cases because re-ipl does not preserve memory and we'd
need a mechanism for Sapphire know that there is a CAPP adapter.

We might be able to reduce this to 750ms or 500ms but need more testing.  Use
1s to be safe.  Also, phyp fw uses 1s after deassert.

Signed-off-by: Ryan Grimm <grimm@linux.vnet.ibm.com>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>