occ: Fix the low level ACK message sent to FSP on receiving {RESET/LOAD}_OCC
Modify the FSP response message to include the status code in the
status/error byte instead of adding a new word to it which is
incorrect.

FSP ack messages are 2 words with status in the 3rd byte of second
word.  Status byte is in the extra (3rd) word only on new status
messages from OPAL to FSP.

Code corrected based on FSP mailbox spec version 3.16.

Signed-off-by: Shilpasri G Bhat <shilpa.bhat@linux.vnet.ibm.com>
Signed-off-by: Vaidyanathan Srinivasan <svaidy@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>