use $(wildcard ) in test Makefiles
The core/test/ and libc/test/ Makefile.check files both contain:

-include core/test/*.d  (or libc/test/*d)

which is incorrect, since that evaluates literally to a *.d file.
This results in each build trying to find that file, and creating
it when not found (in an incorrect way because of other problems in
the Makefile).

The correct way to specify it is:

-include $(wildcard core/test/*.d)

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>