Add Travis configuration
This will allow Travis CI (https://travis-ci.org) to build the firmware
and run the test suite. In the future we can also use it to run and
publish the results of scan-build and lcov.

Travis build environments are stuck back on 12.04, so we don't have a
powerpc cross compiler as part of the archive. Instead download one from
kernel.org and use that.

The 'make check' tests rely on newer -Wstack-usage flags that are not
present on the old version of gcc shipped with 12.04. Install 4.8 from
the toolchain ppa and set it to be the system default.

The tests also require valgrind, so include that in the VM environment.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>