core/device: wrap a function to destroy a dt_node
When destroying a dt_node, it needs to release both the name and itself. And
there are several places to do the release.

This patch wrap a function dt_destroy() to make sure both of the elements of a
dt_node are released when it is not used any more.

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>