ipmi/fru: Remove redundant version information
The FRU data format limits string length to 31 characters. Currently
we truncate the version string as required but leave the redundant
(because it's included in the product name field) "skiboot-" at the
start of the version.

This patch removes the redundant prefix allowing more of the actual
version string to be stored. It also fixes a bug which caused the last
character of a 31 character version string to be overwritten as well
as a rather embarrassing off-by-one buffer overflow.

Note that 31 characters should be enough to store the complete version
string for any release builds.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>