fsp: Avoid NULL dereference in case of invalid class_resp bits
When handling timeouts, we appear to do an occasional NULL dereference
in fsp_timeout_poll() due to fsp_cmdclass_resp_bitmask being out of
sync (bit set but class queue empty). The cause for the discrepancy
will be sorted out separately but the code should be more robust.

Additionally, add a lock to ensure we don't race on the timer calculations
otherwise we might get spurrious dual detection of the timeout.

Fixes SW288484

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>