test/hello_world: Fix some corner cases in test script
The test for expect would always return true. This change makes it do
the correct thing in the presence and absence of expect on my Ubuntu
machine.

Also, skip the test if the user has KERNEL set. With this set skiboot
contains an embedded kernel, and will load that before falling back to
our hello_world test kernel, causing all kinds of confusion.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>