Allow ELF executables with function descriptor entrypoint to be skiboot payloads
Big-endian ELF64 ELF executables normally (the Linux kernel is an
exception) have their entry point refer to a function descriptor instead
of the first instruction. Distinguish between the Linux case and the
function descriptor case, which is used for the FreeBSD kernel, by
checking whether the entry point points into an executable section or
not. This allows use of the FreeBSD kernel as a skiboot payload.

Signed-off-by: Nathan Whitehorn <nwhitehorn@freebsd.org>
[stewart@linux.vnet.ibm.com: reverse test polarity, change continue to break]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>