astbmc: Allow loading of payload from flash
To date BMC platforms have had their payload built in to the skiboot
binary. This patch adds the option of loading from a flash partition
instead.

This has been tested on palmetto with a separate skiboot and
kernel+rootfs partition. In the future we may have separate kernel and
rootfs, which should just work. For now this is what we see when booting:

  [10648940943,5] INIT: Kernel loaded, size: 15728640 bytes (0 = unknown
  preload)
  [10649094578,5] INIT: 32-bit kernel entry at 0x2001015c
  [10649170607,3] PLAT: No ROOTFS partition in PNOR

If the partition named cannot be found, the core loading logic will fall
back to using the built-in payload. In this case, the user gets the
following messages:

  [2214557191,3] PLAT: No KERNEL partition in PNOR.
  [2220486159,5] INIT: platform kernel load failed.
  [2221293286,5] Using built-in kernel.
  [2265861935,5] INIT: Kernel loaded, size: 14706638 bytes (0 = unknown
  preload).

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>